requirements.txt
aiohappyeyeballs==2.4.4 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.11.9 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
altair==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.11" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.11" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
blis==0.7.11 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
catalogue==2.0.10 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
cloudpathlib==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
confection==0.1.5 ; python_version >= "3.11" and python_version < "4.0"
cymem==2.0.10 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
gitdb==4.0.11 ; python_version >= "3.11" and python_version < "4.0"
gitpython==3.1.43 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.11" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.11" and python_version < "4.0"
langchain-core==0.3.21 ; python_version >= "3.11" and python_version < "4.0"
langchain-openai==0.2.10 ; python_version >= "3.11" and python_version < "4.0"
langchain-text-splitters==0.3.2 ; python_version >= "3.11" and python_version < "4.0"
langchain==0.3.9 ; python_version >= "3.11" and python_version < "4.0"
langcodes==3.5.0 ; python_version >= "3.11" and python_version < "4.0"
langfuse==2.55.0 ; python_version >= "3.11" and python_version < "4.0"
langsmith==0.1.147 ; python_version >= "3.11" and python_version < "4.0"
language-data==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
marisa-trie==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
murmurhash==1.0.11 ; python_version >= "3.11" and python_version < "4.0"
narwhals==1.15.2 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
openai==1.56.1 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.10.12 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.0.0 ; python_version >= "3.11" and python_version < "4.0"
preshed==3.0.9 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.2.1 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.0 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==18.1.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.27.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.10.3 ; python_version >= "3.11" and python_version < "4.0"
pydeck==0.9.1 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
rapidfuzz==3.10.1 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.11" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.22.1 ; python_version >= "3.11" and python_version < "4.0"
setuptools==75.6.0 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
smart-open==7.0.5 ; python_version >= "3.11" and python_version < "4.0"
smmap==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
spacy-legacy==3.0.12 ; python_version >= "3.11" and python_version < "4.0"
spacy-loggers==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
spacy-lookups-data==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
spacy==3.7.5 ; python_version >= "3.11" and python_version < "4.0"
spaczz==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.36 ; python_version >= "3.11" and python_version < "4.0"
srsly==2.4.8 ; python_version >= "3.11" and python_version < "4.0"
streamlit-cookies-controller==0.0.4 ; python_version >= "3.11" and python_version < "4.0"
streamlit==1.40.2 ; python_version >= "3.11" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.11" and python_version < "4.0"
thinc==8.2.5 ; python_version >= "3.11" and python_version < "4.0"
tiktoken==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.4.2 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typer==0.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
wasabi==1.1.3 ; python_version >= "3.11" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_system != "Darwin"
weasel==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.18.3 ; python_version >= "3.11" and python_version < "4.0"
